boot.go: split makebootable family into its own file
